Update uList.txt
WARNING!!!

As the repository owner, I would like to provide some advice. It is recommended to avoid using the filter that I created. The filter consists of multiple merged filters, some of which are quite large. Unfortunately, this has led to the accumulation of errors from different sources, false positives, and other issues. It has become impractical to even starting using this complex filter.

Signed-off-by: DrQuackster <128331393+DrQuackster@users.noreply.github.com>